refactor: improve HWAccounts logic and add new accounts
