+++ src/redux/selectors.js
@@ -28,6 +28,7 @@ const selectNetworkName = (state) => state.network.name;
 const selectActiveTokenNetwork = (state) => state.network.networks[state.token.active];
 const selectStaking = (state) => state.staking;
 const selectHWAccounts = (state) => state.hardwareWallet.accounts;
+const selectHWStatus = (state) => state.hardwareWallet.status;
 
 export {
   selectStaking,
@@ -48,4 +49,5 @@ export {
   selectNetworkName,
   selectActiveTokenNetwork,
   selectHWAccounts,
+  selectHWStatus
 };
